+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed
+
+- Fix quick menu edit mode activating on right click analysis menu exit
+
 ## [1.5.5] 2025-11-04
 
 ### Fixed